test_db_connection.py
"""
Test PostgreSQL Supabase Connection
Run: python test_db_connection.py
"""

import os
import sys
from dotenv import load_dotenv
from sqlalchemy import create_engine, text
from sqlalchemy.exc import OperationalError
import traceback

# Load environment variables
load_dotenv()

print("=" * 70)
print("🔍 POSTGRESQL SUPABASE CONNECTION TEST")
print("=" * 70)

# ============================================================
# TEST 1: Check Environment Variables
# ============================================================
print("\n📋 TEST 1: Checking Environment Variables...")

DATABASE_URL = os.getenv('DATABASE_URL')
SUPABASE_URL = os.getenv('SUPABASE_URL')
SUPABASE_ANON_KEY = os.getenv('SUPABASE_ANON_KEY')

if not DATABASE_URL:
    print("❌ DATABASE_URL not found in .env file!")
    print("   Please add DATABASE_URL to your .env file")
    sys.exit(1)
else:
    # Mask password in output
    if '@' in DATABASE_URL:
        masked_url = DATABASE_URL.split('@')[0].split('://')[0] + '://***:***@' + DATABASE_URL.split('@')[1]
        print(f"✅ DATABASE_URL found: {masked_url}")
    else:
        print(f"✅ DATABASE_URL found: {DATABASE_URL[:50]}...")

if SUPABASE_URL:
    print(f"✅ SUPABASE_URL found: {SUPABASE_URL}")
else:
    print("⚠️  SUPABASE_URL not found (optional)")

if SUPABASE_ANON_KEY:
    print(f"✅ SUPABASE_ANON_KEY found: {SUPABASE_ANON_KEY[:30]}...")
else:
    print("⚠️  SUPABASE_ANON_KEY not found (optional)")

# ============================================================
# TEST 2: Parse Database URL
# ============================================================
print("\n🔍 TEST 2: Parsing Database URL...")

try:
    # Extract components
    if '://' in DATABASE_URL:
        protocol = DATABASE_URL.split('://')[0]
        rest = DATABASE_URL.split('://')[1]
        
        if '@' in rest:
            auth = rest.split('@')[0]
            host_db = rest.split('@')[1]
            
            if ':' in auth:
                username = auth.split(':')[0]
                password_part = auth.split(':')[1]
            else:
                username = auth
                password_part = ''
            
            if '/' in host_db:
                host_port = host_db.split('/')[0]
                database = host_db.split('/')[1].split('?')[0]
            else:
                host_port = host_db
                database = 'postgres'
            
            if ':' in host_port:
                host = host_port.split(':')[0]
                port = host_port.split(':')[1]
            else:
                host = host_port
                port = '5432'
            
            print(f"✅ Protocol: {protocol}")
            print(f"✅ Username: {username}")
            print(f"✅ Password: {'*' * 10}")
            print(f"✅ Host: {host}")
            print(f"✅ Port: {port}")
            print(f"✅ Database: {database}")
            
            # Check SSL parameters
            if '?' in DATABASE_URL:
                params = DATABASE_URL.split('?')[1]
                print(f"✅ Parameters: {params}")
        else:
            print("⚠️  No authentication found in URL")
    else:
        print("⚠️  Invalid URL format")
        
except Exception as e:
    print(f"⚠️  Error parsing URL: {str(e)}")

# ============================================================
# TEST 3: Create SQLAlchemy Engine
# ============================================================
print("\n🔧 TEST 3: Creating SQLAlchemy Engine...")

try:
    engine = create_engine(
        DATABASE_URL,
        pool_size=5,
        pool_pre_ping=True,
        echo=False
    )
    print("✅ SQLAlchemy engine created successfully")
except Exception as e:
    print(f"❌ Error creating engine: {str(e)}")
    traceback.print_exc()
    sys.exit(1)

# ============================================================
# TEST 4: Test Connection
# ============================================================
print("\n🔌 TEST 4: Testing Database Connection...")

try:
    with engine.connect() as connection:
        print("✅ Connection established!")
        
        # Test query
        result = connection.execute(text("SELECT version();"))
        version = result.fetchone()[0]
        print(f"✅ PostgreSQL version: {version[:50]}...")
        
except OperationalError as e:
    print(f"❌ Connection failed!")
    print(f"   Error: {str(e)}")
    print("\n🔍 Common issues:")
    print("   1. Check if Supabase project is ACTIVE (not paused)")
    print("   2. Verify password is correct")
    print("   3. Check firewall/antivirus blocking port 6543")
    print("   4. Try adding '?sslmode=require' to DATABASE_URL")
    traceback.print_exc()
    sys.exit(1)
except Exception as e:
    print(f"❌ Unexpected error: {str(e)}")
    traceback.print_exc()
    sys.exit(1)

# ============================================================
# TEST 5: Check Tables
# ============================================================
print("\n📊 TEST 5: Checking Database Tables...")

try:
    with engine.connect() as connection:
        # Query for tables in public schema
        result = connection.execute(text("""
            SELECT table_name 
            FROM information_schema.tables 
            WHERE table_schema = 'public' 
            ORDER BY table_name;
        """))
        
        tables = [row[0] for row in result.fetchall()]
        
        if tables:
            print(f"✅ Found {len(tables)} table(s):")
            for table in tables:
                print(f"   - {table}")
        else:
            print("⚠️  No tables found in database")
            print("   This is normal for a new database")
            print("   Tables will be created when you run app.py")
        
except Exception as e:
    print(f"⚠️  Error checking tables: {str(e)}")

# ============================================================
# TEST 6: Test Write Permission
# ============================================================
print("\n✍️  TEST 6: Testing Write Permission...")

try:
    with engine.connect() as connection:
        # Create a temporary test table
        connection.execute(text("""
            CREATE TABLE IF NOT EXISTS _test_connection (
                id SERIAL PRIMARY KEY,
                test_value TEXT,
                created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
            );
        """))
        connection.commit()
        print("✅ CREATE TABLE successful")
        
        # Insert test data
        connection.execute(text("""
            INSERT INTO _test_connection (test_value) 
            VALUES ('Connection test successful');
        """))
        connection.commit()
        print("✅ INSERT successful")
        
        # Read test data
        result = connection.execute(text("""
            SELECT test_value FROM _test_connection LIMIT 1;
        """))
        test_value = result.fetchone()[0]
        print(f"✅ SELECT successful: '{test_value}'")
        
        # Clean up
        connection.execute(text("DROP TABLE _test_connection;"))
        connection.commit()
        print("✅ DROP TABLE successful")
        
        print("✅ All write operations successful!")
        
except Exception as e:
    print(f"❌ Write test failed: {str(e)}")
    print("   Check database permissions")
    traceback.print_exc()

# ============================================================
# TEST 7: Connection Pool Test
# ============================================================
print("\n🏊 TEST 7: Testing Connection Pool...")

try:
    # Open multiple connections
    connections = []
    for i in range(3):
        conn = engine.connect()
        connections.append(conn)
        print(f"✅ Connection {i+1}/3 opened")
    
    # Close all connections
    for i, conn in enumerate(connections):
        conn.close()
        print(f"✅ Connection {i+1}/3 closed")
    
    print("✅ Connection pooling working correctly")
    
except Exception as e:
    print(f"⚠️  Connection pool test failed: {str(e)}")

# ============================================================
# FINAL SUMMARY
# ============================================================
print("\n" + "=" * 70)
print("📊 TEST SUMMARY")
print("=" * 70)

print("""
✅ Environment Variables: OK
✅ Database URL Parsing: OK
✅ SQLAlchemy Engine: OK
✅ Database Connection: OK
✅ Table Access: OK
✅ Write Permissions: OK
✅ Connection Pooling: OK

🎉 ALL TESTS PASSED!

Your PostgreSQL Supabase connection is working perfectly.
You can now run: python app.py
""")

print("=" * 70)