requirements.txt
# Core Framework
Flask
Werkzeug

# Database
Flask-SQLAlchemy
psycopg2-binary

# Data Processing
pandas
numpy
plotly
openpyxl

# ML Inference (ONNX Runtime only - no training libraries)
onnxruntime==1.16.3
onnxmltools==1.12.0

# Utilities
python-dotenv
requests
Pillow
python-dateutil
Flask-WTF
WTForms

# Security & Auth
Flask-Login
email-validator
bcrypt
psutil
gunicorn
